% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSAS.create.transects.R
\name{ECSAS.add.sample.label}
\alias{ECSAS.add.sample.label}
\title{Add sample labels to a dataframe of watches}
\usage{
ECSAS.add.sample.label(wtchs, trns)
}
\arguments{
\item{wtchs}{a dataframe of watches containing at least a column name WatchID}

\item{trns}{a transect object as returned by \code{ECSAS.create.transects()}.}
}
\description{
Add a \code{Sample.Label} column to \code{wtchs} and set it to the
corresponding \code{Sample.Label} from \code{trns} that each watch belongs to.
}
\details{
Simpy unnests the \code{Watches} list-column from \code{trns} and joins
it to the \code{wtchs} dataframe matching on \code{WatchID}. Replaces existing
\code{Sample.Label} column if it exists. Throws an error if any row of \code{trns}
has no associated watches.
}
