% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSAS.create.aerial.watches.R
\name{ECSAS.create.aerial.watches}
\alias{ECSAS.create.aerial.watches}
\title{Convert a series of continuous observations into segments}
\usage{
ECSAS.create.aerial.watches(dat, watchlen = 30, posns = NULL, ecsas.path, ...)
}
\arguments{
\item{dat}{Input data frame of continuous observations. Can contain data for
multiple surveys. Must contain a transect column whose name is provided
in \code{transcol}}

\item{watchlen}{The length in seconds of each segment to be created (default 30
seconds).}

\item{posns}{The GPS track positions with one row per second.}

\item{ecsas.path}{[character, default: \sQuote{NULL}]\cr Full path name to
the ECSAS database. Needed to extract the pause/resume info for aerial data.}
}
\value{
a dataframe of watches
}
\description{
This function takes a data frame of continuous observations and
converts them into the segments Each row of the \code{dat} data frame
contains either an observation, a PAUSE, or a RESUME record, Each
transect is cut into segments (of \code{seglen} seconds each) beginning
from the time of the transect start record. Thus the final watch in each
transect may not be a full \code{seglen} seconds long.
}
\details{
This is mostly used to convert aerial survey transects into segments
for Density Surface Modeling and other analyses.
}
\section{Author}{
Dave Fifield
}

