% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSAS.find.suspicious.dist.R
\name{ECSAS.find.suspicious.dist}
\alias{ECSAS.find.suspicious.dist}
\title{Find ECSAS watches with suspicious distances}
\usage{
ECSAS.find.suspicious.dist(dat = NULL, thresh = 50, debug = FALSE)
}
\arguments{
\item{dat}{a dataframe, such as returned by \code{\link{ECSAS.extract()}}, containing, at a minimum, fields
named \code{LatStart, LatEnd, LongStart, LongEnd, PlatformSpeed, CalcDurMin.}}

\item{thresh}{(as percentage; default 50) maximum percentage of the smaller of the two distances that the
difference is allowed to be before being flagged as problematic.}
}
\value{
Returns \code{dat} with the following addtional columns added:

\tabular{ll}{
\code{dist_geo_km} \tab distance as calculated by \code{\link[geosphere:distGeo]{geosphere::distGeo()}} \cr
\code{dist_dr_km} \tab distance as calculated by dead reckoning: \eqn{PlatformSpeed * (CalcDurMin / 60) * 1.852}.
Note the conversion from nautical miles to km.\cr
\code{dist_diff_km} \tab difference in the two distances \cr
\code{dist_pct_diff} \tab percentage difference (of the smaller distance) \cr
}

Users should look closely at watches with large \code{pct_diff} AND/OR large \code{dist_diff_km}.
}
\description{
Find watches where the distance traveled as calculated by lat/long vs dead reckoning seems suspicious.
This function looks for watches whose length in km as calculated by dead-reckoning (speed * time)
is \emph{significantly} different than that calculated by using \code{\link[geosphere:distGeo]{geosphere::distGeo()}} from start and end
positions. A watch is flagged as significantly different if the difference between the two distance
measures is \code{>= thresh} percent of the smaller of the two distances.
}
\section{Author}{
Dave Fifield
}

\seealso{
\code{\link{ECSAS.extract()}}
}
