% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSAS.create.transects.R
\name{ECSAS.find.trans}
\alias{ECSAS.find.trans}
\title{Find transect that each watch belongs to}
\usage{
ECSAS.find.trans(trns, watchid)
}
\arguments{
\item{trns}{a transect object as returned by \code{ECSAS.create.transects}. dataframe of ECSAS watches (1 row per watch).}

\item{watchid}{a vector of watchid's}
}
\description{
This function will return the transect id each watch is a vector of watchid's.
}
\details{
Simpy looks for any transects in \code{tns} that include \code{watchid} in its
\code{Watches} column.
}
