% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSAS.get.cruises.R
\name{ECSAS.get.cruises}
\alias{ECSAS.get.cruises}
\title{Get ECSAS cruises  from database}
\usage{
ECSAS.get.cruises(ecsas.path = NULL, debug = FALSE)
}
\arguments{
\item{ecsas.path}{(required) full path name to the ECSAS database.}
}
\description{
This function will connect to the Access database and return a
data frame containing info about all cruises.
}
\section{Author}{
Dave Fifield
}

