% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSAS.get.table.R
\name{ECSAS.get.table}
\alias{ECSAS.get.table}
\title{Get ECSAS table from database}
\usage{
ECSAS.get.table(ecsas.path = NULL, table = NULL, debug = FALSE)
}
\arguments{
\item{ecsas.path}{(default NULL) full path name to the ECSAS database.}

\item{table}{character vector (of length 1) naming table to be returned.}
}
\description{
This function will connect to the Access database and return a table as a data frame.
}
\section{Author}{
Dave Fifield
}

