% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSAS.run.query.R
\name{ECSAS.run.query}
\alias{ECSAS.run.query}
\title{Execute an SQL query on an ECSAS database}
\usage{
ECSAS.run.query(ecsas.path, sqlstr)
}
\arguments{
\item{ecsas.path}{[character]\cr Full path name to the ECSAS
database.}

\item{strsql}{SQL string to execute}
}
\value{
Returns whatever the query returned.
}
\description{
This function will connect to the Access database, execute a query and
return the result.
}
\details{
Opens a connection to the database and executes the query given in \code{sqlstr}.
}
\section{Author}{
Dave Fifield
}

