\name{ECSASnames}
\alias{ECSASnames}
\docType{data}
\title{
Names for exports to ECSAS
}
\description{
This \code{data.frame} gives the order of columns and names for an export to ECSAS.
}
\usage{data("ECSASnames")}
\format{
  A data frame with several columns.
}
\details{
 For internal use only.
}
\source{
}
\references{
}
\examples{
}
\keyword{datasets}
