% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tblmissionselect.delete.R
\name{tblmissionselect.delete}
\alias{tblmissionselect.delete}
\title{MAke sute that the table tblmissionselect doesnt exist in the ECSAS database}
\usage{

  tblmissionselect.delete(ecsas.drive = "C:/Users/Christian/Google Drive/ECSAS/",
  ecsas.file = "Oiseaux marins 2006-2014.accdb")
}
\arguments{
\item{ecsas.drive}{Where is located the ECSAS Access database}

\item{ecsas.file}{What is the name of the ECSAS Access database}
}
\description{
Sometimes the table tblmissionselect is not properly deleted from the ECSAS database after the extraction. This function make sure that the file is properly deleted.
}
\details{
Delete the table tblmissionselect form the ECSAS database
}
\section{Author}{
Christian Roy
}

