% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tblspselect.delete.R
\name{tblspselect.delete}
\alias{tblspselect.delete}
\title{MAke sute that the table tblspselect doesnt exist in the ECSAS database}
\usage{
tblspselect.delete(ecsas.drive = "C:/Users/Christian/Google Drive/ECSAS/",
  ecsas.file = "Master ECSAS v 3.33.mdb")
}
\arguments{
\item{ecsas.drive}{Where is located the ECSAS Access database}

\item{ecsas.file}{What is the name of the ECSAS Access database}
}
\description{
Sometimes the table tblspselect is not properly deleted from the ECSAS database after the extraction. This function make sure that the file is properly deleted.
}
\details{
Delete the table tblspselect form the ECSAS database
}
\section{Author}{
Christian Roy
}

