% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSAS.create.aerial.watches.R
\name{ECSAS.create.aerial.watches}
\alias{ECSAS.create.aerial.watches}
\title{Create watches from aerial transects}
\usage{
ECSAS.create.aerial.watches(
  dat,
  watchlen = 30,
  posns = NULL,
  ecsas.path,
  verbose = FALSE
)
}
\arguments{
\item{dat}{Input data frame of transect info. Can contain data for
multiple surveys. Must contain at least columns for transect ID
(\code{AerialTransectID}), and transect start/end date and time
(\code{StartDateTime}, \code{EndDateTime}). All other columns are ignored.
These columns need not be unique between rows (eg. if \code{dat} contains the
actual animal observations). Watches will only be created for
each unique transect.}

\item{watchlen}{The length in seconds of each watch to be created (default 30
seconds).}

\item{posns}{The GPS track positions with one row per second. \code{posns} must
contain columns \code{datetime, lat, long}, and \code{alt}, the latter of which
may contain \code{NA}.}

\item{ecsas.path}{Full path name to the ECSAS database. Needed to extract the
pause/resume info for aerial data. Default \code{NULL}.}

\item{verbose}{If \code{TRUE}, the \code{AerialTransectID} of each transect
is printed as it is processed. Default \code{FALSE}.}
}
\value{
A dataframe of watches.
}
\description{
This function converts aerial transect data into watches
(ie segments).
}
\details{
This is mostly used to convert aerial survey transects into segments
for Density Surface Modeling and other analyses. Periods of off-effort during
transects (e.g., while passing over islands) will be ignored. Off-effort info
is extracted directly from the database. Start and end positions of each
watch are assigned from \code{posns}, if supplied, based on matching time.
A warning is issued if any start/end positions cannot be assigned.

The final watch in each transect (or before an off-effort break) may not be
a full \code{watchlen} seconds long.
}
\section{Author}{
Dave Fifield
}

