% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSAS.extract.R
\name{ECSAS.extract}
\alias{ECSAS.extract}
\title{Extract data from ECSAS database}
\usage{
ECSAS.extract(
  species,
  years,
  lat = c(-90, 90),
  long = c(-180, 180),
  obs.keep = NA,
  obs.exclude = NA,
  sub.program = c("All", "Atlantic", "Quebec", "Arctic", "ESRF", "AZMP", "FSRS"),
  intransect = TRUE,
  distMeth = c(14, 20),
  ecsas.path = NULL,
  ecsas.drive = "C:/Users/christian/Dropbox/ECSAS",
  ecsas.file = "Master ECSAS_backend v 3.31.mdb",
  debug = FALSE
)
}
\arguments{
\item{species}{Optional. Alpha code (or vector of Alpha codes, e.g., c("COMU,"TBMU", "UNMU")) for the species desired in the extraction.}

\item{years}{Optional. Either a single year or a vector of two years denoting "from" and "to" (inclusive).}

\item{lat}{Pair of coordinate giving the southern and northern limits of the range desired.}

\item{long}{Pair of coordinate giving the western and eastern limits of the range desired. Note that west longitude values must be negative.}

\item{obs.keep}{Name of the observer to keep for the extraction. The name of the observer must be followed by it's first name (eg: "Bolduc_Francois").}

\item{obs.exclude}{Name of the observer to exlude for the extraction.The name of the observer must be followed by it's first name (eg: "Bolduc_Francois").}

\item{sub.program}{From which sub.program the extraction must be made. Options are Quebec, Atlantic, Arctic, ESRF, AZMP, FSES, or All
All subprograms will inlcude the observations made in the PIROP program.}

\item{intransect}{Should we keep only the birds counted on the transect (if TRUE, the default) or extract all observations (if FALSE).}

\item{distMeth}{Integer specifying the distance sampling method code (tblWatch.DistMeth in ECSAS). Default is c(14, 20) which includes all watches
with perpendicular distanes for both flying and swimming birds. If "All", then observations from all distance sampling methods will be returned.}

\item{ecsas.path}{(default NULL) full pathname to the ECSAS database. If NULL, the path is built from \code{ecsas.drive} and \code{ecsas.file}.}

\item{ecsas.drive}{path to folder containing the ECSAS Access database}

\item{ecsas.file}{name of the ECSAS Access database}
}
\description{
The function will connect to the Access database, create a series of queries and import the desired information in a data frame.
}
\details{
The function will produce a data frame that contains all the pertinent information. Note that watches with no observations (the so called "zeros" are
included by default).

The distance traveled during the watch is returned in the column \code{WatchLenKm}. If lat/long coordinates
are available for both the start and end locations of the watch, then it is calculated as the shortest 
distance between these two points on the WGS84 ellipsoid using \code{geosphere::DistGeo} and, in this case,
\code{WatchLenKmHow} will contain \code{"distGeo"}. Otherwise \code{WatchLenKm} 
is calculated as the \code{PlatformSpeed * CalcDurMin} where \code{CalcDurMin} is the length of the 
watch in minutes computed from start and end times. In this case, \code{WatchLenKmHow} will contain 
\code{"Dead Reckoning"}.
}
\section{Author}{
Christian Roy, Dave Fifield
}

\seealso{
\code{\link{QC.extract}}
}
