% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ECSAS.extract.R
\name{ECSAS.extract}
\alias{ECSAS.extract}
\title{Extract the information for the Global ECSAS database}
\usage{
ECSAS.extract(sp = "ATPU", years = c(2006, 2013), lat = c(30, 70),
  long = c(-70, -30), Obs.keep = NA, Obs.exclude = NA,
  database = c("Atlantic", "Quebec", "Both"), snapshot = T,
  intransect = T, ecsas.drive = "C:/Users/christian/Dropbox/ECSAS",
  ecsas.file = "Master ECSAS_backend v 3.31.mdb")
}
\arguments{
\item{sp}{Alpha code for the specie desired in the extraction. If more than one species all the desired species must be saved into a vector ex: c("COMU,"TBMU", "UNMU")}

\item{years}{Years desired for the extraction}

\item{lat}{Pairs of coordinate giving the southern and northern limits of the range desired.}

\item{long}{Pairs of coordinate giving the western and eastern limits of the range desired. Note that longitude values must be negative.}

\item{Obs.exclude}{Name of the observer to exlude for the extraction.The name of the observer must be followed by it's first name (eg: "Bolduc_Francois").}

\item{database}{From which database the extraction must be made. Currently restricted to Atlantic and/or Quebec.}

\item{snapshot}{Should we keep only the birds counted during the snapshot or not}

\item{intransect}{Should we keep only the birds counted on the transect or not.}

\item{ecsas.drive}{Where is located the ECSAS Access database}

\item{ecsas.file}{What is the name of the ECSAS Access database}

\item{obs.keep}{Name of the observer to keep for the extraction. The name of the observer must be followed by it's first name (eg: "Bolduc_Francois").}
}
\description{
The function will connect to the Access database, create a series of queries and import the desired information in a data frame.
}
\details{
The function will produce a data frame that will contains all the pertinent information.
}
\section{Author}{
Christian Roy
}
\seealso{
\code{\link{QC.extract}}
}

