% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSAS.extract.R
\name{ECSAS.extract}
\alias{ECSAS.extract}
\title{Extract data from ECSAS database}
\usage{
ECSAS.extract(
  species,
  years,
  lat = c(-90, 90),
  long = c(-180, 180),
  obs.keep = NA,
  obs.exclude = NA,
  sub.program = c("All", "Atlantic", "Quebec", "Arctic", "ESRF", "AZMP", "FSRS"),
  intransect = TRUE,
  distMeth = c(14, 20),
  ecsas.path = NULL,
  ecsas.drive = "C:/Users/christian/Dropbox/ECSAS",
  ecsas.file = "Master ECSAS_backend v 3.31.mdb",
  ind.tables.only = FALSE,
  debug = FALSE
)
}
\arguments{
\item{species}{Optional. Alpha code (or vector of Alpha codes, e.g., c("COMU,"TBMU", "UNMU")) for the species desired in the extraction.}

\item{years}{Optional. Either a single year or a vector of two years denoting "from" and "to" (inclusive).}

\item{lat}{Vector of two numbers giving the southern and northern limits of the range desired.}

\item{long}{Vector of two numbers giving the western and eastern limits of the range desired. Note that west longitude values must be negative.}

\item{obs.keep}{Names of the observers to keep for the extraction. Name format: Surname_FirstName (eg: "Bolduc_Francois").}

\item{obs.exclude}{Name of the observer to exclude for the extraction. Name format: Surname_FirstName (eg: "Bolduc_Francois").}

\item{sub.program}{From which sub.program the extraction must be made. Options are Quebec, Atlantic,
Arctic, ESRF, AZMP, FSES, or All.
All subprograms will include the observations made in the PIROP program.}

\item{intransect}{If TRUE (the default), return only observations coded as "In Transect", otherwise return all
observations. See the ECSAS survey protocol for more details:

Gjerdrum, C., D.A. Fifield, and S.I. Wilhelm.
2012. Eastern Canada Seabirds at Sea (ECSAS) standardized protocol for pelagic seabird surveys from
moving and stationary platforms. Canadian Wildlife Service Technical Report Series No. 515.
Atlantic Region. vi + 37 pp.}

\item{distMeth}{Integer(s) specifying the distance sampling method code(s) (see tblWatch.DistMeth in ECSAS).
Acceptable values are a single integer, a vector of integers, or "All".
Default is c(14, 20) which includes all watches with perpendicular distances for both flying and swimming birds.
If "All", then observations from all distance sampling methods will be returned, which may include
observations from the PIROP program if no other options preclude this.}

\item{ecsas.path}{(default NULL) full path name to the ECSAS database. If NULL, the path is built from
\code{ecsas.drive} and \code{ecsas.file}.}

\item{ecsas.drive}{path to folder containing the ECSAS Access database}

\item{ecsas.file}{name of the ECSAS Access database file}

\item{ind.tables.only}{Indicates if two individual tables for watch/cruise, and observations
should be returned (default FALSE) rather than a single table with all columns combined.
See Value section.}
}
\value{
By default the function will produce a data frame that contains all the pertinent information.
Note that watches with no observations (the so called "zeros" are included by default).

If \code{ind.tables.only} is \code{FALSE} (the default), then a single dataframe is returned
containing all pertinent cruise, watch and sightings table info for each observation. If a given watch had
no observations, then sighting related fields will be \code{NA}.

If \code{ind.tables.only} is \code{TRUE}, then a list is returned with the following elements:

\tabular{ll}{
\code{watches} \tab the combined columns from the watch and cruise tables only.\cr
\code{sightings}\tab the columns from the sightings table only.
}

Note it is not currently possible to extract the watch and cruise tables separately.
}
\description{
This function will connect to the Access database, create a series of queries and import the desired information in a data frame.
}
\details{
The distance traveled during the watch is returned in the column \code{WatchLenKm}. If lat/long coordinates
are available for both the start and end locations of the watch, then it is calculated as the shortest
distance between these two points on the WGS84 ellipsoid using \code{\link[geosphere:distGeo]{geosphere::distGeo()}} and, in this case,
\code{WatchLenKmHow} will contain \code{"distGeo"}. Otherwise \code{WatchLenKm}
is calculated as the \code{PlatformSpeed * CalcDurMin} where \code{CalcDurMin} is the length of the
watch in minutes computed from start and end times. In this case, \code{WatchLenKmHow} will contain
\code{"Dead Reckoning"}.
}
\section{Author}{
Christian Roy, Dave Fifield
}

\seealso{
\code{\link{QC.extract}}
}
